Rewrite how data is collected and sent.
Based on feedback/discussions with Nicu Pavel the parser should be
sending accumulated average or accumulated total values, not hourly.
Only the last reported value for the day is saved.

Providing data this way actually simplifies the program. We simply
keep a running average/total in the UDP listener and report out the
current value at each parseInterval.

Note that I'm not convinced that a running average makes sense for
solar radiation (or wind speed).